+++ packages/design-components/scripts/fetchColor.js
@@ -2,26 +2,10 @@ import fs from 'node:fs'
 import path from 'node:path'
 
 import {setColor} from '@ndive/design-tokens'
-import * as dotenv from 'dotenv'
 
-function normalize(relativePath) {
-    return path.resolve(import.meta.dirname, relativePath)
-}
-
-dotenv.config({path: normalize('../.env')})
-
-function snakeToPascal(snake) {
-    return snake.replace(/(_\w)/g, function (match) {
-        return match[1].toUpperCase()
-    })
-}
-
-function pascalToCamel(str) {
-    if (str.length === 0) {
-        return str
-    }
-    return str.charAt(0).toLowerCase() + str.slice(1)
-}
+import './utils/loadEnv.js'
+import {normalize} from './utils/normalize.js'
+import {pascalToCamel, snakeToPascal} from './utils/string.js'
 
 async function fetchColor() {
     const {group, colorSet} = await setColor({
@@ -29,7 +13,7 @@ async function fetchColor() {
     })
 
     // color.scss
-    const stylesDir = normalize('../src/styles')
+    const stylesDir = normalize(import.meta.dirname, '../src/styles')
     !fs.existsSync(stylesDir) &&
         fs.mkdirSync(stylesDir, {
             recursive: true,
@@ -47,7 +31,7 @@ async function fetchColor() {
     })
 
     // color.json
-    const jsonDir = normalize('../src/json')
+    const jsonDir = normalize(import.meta.dirname, '../src/json')
 
     !fs.existsSync(jsonDir) &&
         fs.mkdirSync(jsonDir, {
@@ -65,7 +49,7 @@ async function fetchColor() {
         .join(' | ')};
 `
 
-    fs.writeFileSync(path.join(normalize('../src/types'), 'color.types.ts'), colorTypeDefinition, {
+    fs.writeFileSync(path.join(normalize(import.meta.dirname, '../src/types'), 'color.types.ts'), colorTypeDefinition, {
         encoding: 'utf-8',
     })
 }
