+++ packages/design-components/scripts/utils/loadEnv.js
@@ -1,0 +1,5 @@
+import * as dotenv from 'dotenv'
+
+import {normalize} from './normalize.js'
+
+dotenv.config({path: normalize(import.meta.dirname, '../../.env')})
