+++ packages/design-components/scripts/utils/normalize.js
@@ -1,0 +1,5 @@
+import path from 'node:path'
+
+export function normalize(dirname, relativePath) {
+    return path.resolve(dirname, relativePath)
+}
